+++ src/bittytax/conv/parsers/scripts/kraken_assets.py
@@ -2,11 +2,12 @@
 # (c) Nano Nano Ltd 2020
 # Generate the constants for the Kraken parser and verify the split method
 
-from typing import Dict, List
+from typing import Dict, List, Tuple
 
 import requests
+from colorama import Back
 
-from bittytax.conv.parsers.kraken import _split_trading_pair
+from bittytax.conv.parsers.kraken import ALT_ASSETS, QUOTE_ASSETS, _split_trading_pair
 
 
 def get_alt_assets() -> Dict[str, str]:
@@ -24,10 +25,11 @@ def get_alt_assets() -> Dict[str, str]:
     return alt_assets
 
 
-def get_quote_assets() -> List[str]:
+def get_quote_assets() -> Tuple[List[str], Dict[str, str]]:
     response = requests.get("https://api.kraken.com/0/public/AssetPairs", timeout=10)
 
     quote_assets = []
+    specials = {}
     if response:
         passed = True
         for pair in response.json()["result"]:
@@ -52,6 +54,7 @@ def get_quote_assets() -> List[str]:
                 elif bt_base == base and bt_quote == quote:
                     print(f"{pair} = {bt_base}/{bt_quote} [OK]")
                 else:
+                    specials[pair] = wsname.split("/")[1]
                     passed = False
                     print(f"{pair} = {bt_base}/{bt_quote} [Failure] {wsname} ({base} & {quote})")
 
@@ -62,20 +65,33 @@ def get_quote_assets() -> List[str]:
     else:
         print(f"{response.status_code} {response.reason}")
 
-    return quote_assets
+    return quote_assets, specials
 
 
-def output_constants(alt_assets: Dict[str, str], quote_assets: List[str]) -> None:
+def output_constants(
+    alt_assets: Dict[str, str], quote_assets: List[str], specials: Dict[str, str]
+) -> None:
     print("\nQUOTE_ASSETS = [")
     for i in sorted(quote_assets):
-        print(f'    "{i}"')
+        if i in QUOTE_ASSETS:
+            print(f'    "{i}"')
+        else:
+            print(f'    {Back.RED}"{i}"{Back.RESET}')
     print("]")
 
     print("\nALT_ASSETS = {")
     for k, v in sorted(alt_assets.items()):
-        print(f'    "{k}": "{v}",')
+        if k in ALT_ASSETS:
+            print(f'    "{k}": "{v}",')
+        else:
+            print(f'    {Back.RED}"{k}": "{v}"{Back.RESET},')
+    print("}")
+
+    print("\nTRADINGPAIR_TO_QUOTE_ASSET = {")
+    for i in sorted(specials):
+        print(f'    {Back.RED}"{i}": "{specials[i]}"{Back.RESET},')
     print("}")
 
 
 if __name__ == "__main__":
-    output_constants(get_alt_assets(), get_quote_assets())
+    output_constants(get_alt_assets(), *get_quote_assets())